requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pdfkit==0.6.1
Pillow==8.0.1
psutil==5.7.3
pycparser==2.20
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
wkhtmltopdf==0.2
WTForms==2.3.3
